Move bulk traits to traits folder
