+++ src/Zendesk/API/Resources/JobStatuses.php
@@ -2,14 +2,14 @@
 
 namespace Zendesk\API\Resources;
 
-use Zendesk\API\BulkTraits\BulkFindTrait;
 use Zendesk\API\Traits\Resource\Find;
+use Zendesk\API\Traits\Resource\FindMany;
 
 class JobStatuses extends ResourceAbstract
 {
     const OBJ_NAME = 'job_status';
     const OBJ_NAME_PLURAL = 'job_statuses';
 
-    use BulkFindTrait;
+    use FindMany;
     use Find;
 }
